Fix sorted aggregator factory
